+++ registry/lib/pages/pages.dart
@@ -5,3 +5,4 @@ export 'input_page.dart';
 export 'checkbox_page.dart';
 export 'dropdown_page.dart';
 export 'switch_page.dart';
+export 'radio_page.dart';